requirements.txt

streamlit>=1.35.0
pytest>=8.0.0
PyYAML>=6.0

# Optional LLM providers (only needed when enabling LLM_API_KEY flows)
openai>=1.30.0
anthropic>=0.25.0
